104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,22 +10,22 @@
 int main(void)
 {
 	int i = 3;
-	unsigned long int first = 1;
-	unsigned long int second = 2;
-	unsigned long int next = first + second;
+	long long int first = 1;
+	long long int second = 2;
+	long long int next = first + second;
 
-	printf("%lu, ", first);
-	printf("%lu, ", second);
+	printf("%lld, ", first);
+	printf("%lld, ", second);
 
 	while (i <= 98)
 	{
 		if (i == 98)
 		{
-			printf("%lu\n", next);
+			printf("%lld\n", next);
 		}
 		else
 		{
-			printf("%lu, ", next);
+			printf("%lld, ", next);
 		}
 
 		first = second;
